WIP - room user profile
